New format specifications: meta.db (+ friends) and metrics.yaml
meta.db is the binary replacement for experiment.xml ("database metadata"), designed to be:
 - Random-access: all structures are expandable but of known size,
 - Mmap-friendly: most fields are aligned (on x86) and all are little-endian,
 - Lazily-loadable: all structures can be "skipped" without breaking loading of other structures,
 - and most importantly, properly match the Context framework built up in Prof2.

To keep pieces consistent, the other *.db formats have been updated to follow these guidelines.

metrics.yaml is a forthcoming extension to support taxonomic metrics, this will allow proper
presentation and analysis of dozens of metrics. This is done in YAML to support user-written
metric specifications (ie. upgraded derived metrics).

metrics.yaml currently has no users, making it a perfect format with no flaws. This is the initial
specification that folds a number of ideas together, it will need to be reviewed and revised once
the Viewer adds support. Adding it now since it's a requirement for meta.db.