Tune the static TLS block size when using LD_AUDIT
Glibc has a bug where the static TLS block is not allocated large enough when
LD_AUDIT is specified, causing an uncatchable crash before the application starts.
This commit temporarily avoids the issue by setting GLIBC_TUNABLES to "tune"
the "surplus" added to the block to 64KB (from 1664B).
An --auditor-static-tls-size option is added to hpcrun to control this size.

As a side effect of this, the memory usage per thread increases by 63KB (by default),
however as the additional pages are not touched there is little performance hit on x86-64.