Unwind callers not returns
When unwinding a stack frame, the PC in the resulting frame is a return address, rather
than a call address. If the compiler did not expect the call to return, the compiler may
pack instructions in a way such that the unwind recipe for the return address is not the
same as the call address.

This commit changes the semantic of hpcrun_unw_step to always result in a "caller" frame,
with an approximate call address of (return address - 1), instead of the "return" frame
returned previously. This simplifies the incomplete logic added in b99930f9b3d4bb3d9e384e9f91e212cc0df53d53
and extends it to apply to the generic libunwind and Power cases as well.

The x86 native unwinder is unchanged in this commit, it fails to unwind even in the
microtest case used to validate this patch.