temporarily disable OpenCL support
OpenCL needs work. Currently, ND range kernels need to compute the
module id. That only happens when gtpin is used. This is a mess
and needs major updating. One should be able to measure OpenCL
programs without using GTPin. The change where getModuleId is
needed by the OpenCL API came from the intel_opencl_analysis
branch and is not present in master. whatever ...