fix ompt support for target nowait
when a target nowait is executed outside an explicit parallel
region, it is nested only in the implicit (serialized) parallel
region surrounding the initial task. the implicit parallel
region does not get a parallel begin callback, so its
parallel_data is uninitialized. in this case, the nowait task
can't inherit its callpath from the enclosing parallel region
and it needs to unwind the call stack to determine its call path.