Avoid adding vmlinux load module every time perf init is called
perf init is called every time a thread "starts".
I wrongly assumed perf_init is called only once, but in fact it is called every
time a thread is initialized. This doesn't work if the executable forks multiple
times.
If the initialization is done in the declaration, there is no need to reset
again in perf_init.