hpcstruct: refine use of OMP_STACKSIZE
in order of priority
- consult '-s' option with which user can specify size
- use value of OMP_STACKSIZE (if present)
- set OMP_STACKSIZE=8M

Note: hpcstruct now processes the '-s' option in the shell
script and the rest of the options in the program. this is
an unholy mix. however, setting OMP_STACKSIZE in the
application with putenv doesn't work. Perhaps OMP_STACKSIZE
is inspected by a library constructor. this will due for now
as a hot fix.