hpcrun: Fix PAPI issue with forks
The reason why hpcrun "crashed" when profiling forked executables with
PAPI is because in the child process, when hpcrun calls
PAPI_thread_init, PAPI library is not initialized although it's
initialized in the parent process.

The solution is to call PAPI_library_init before calling
PAPI_thread_init when a thread is created.