Rework the start-up sequence for the sample sources.  This fixes the fork + papi problem where the PAPI source would fail and/or segfault in a process that fork()s.  The key points are:
  1. If a process forks, stop the samples only in the thread that
  calls fork, and restart them after the fork.  Don't shutdown the
  source in the parent since that would kill samples in all threads.

  2. In the child after a fork, shutdown the sources and restart them
  from scratch.  The child has one thread with a new pid, so this is
  the safest way to start the samples.

  3. Treat every process as threaded and always call the thread-init
  method after library-init.  This doesn't cause any trouble and there
  are too many code paths to call thread-init selectively.

Note: we call library-init on all sources, whether used or not.  (PAPI
requires this.)  But we call thread-init only on the sources that are
used.  So, the sequence must be: library-init, process-event-list (to
determine which sources are used), and then thread-init.

Cleanup a couple of -dd messages.  Collapse PRE_FORK and POST_FORK
into just FORK.  There are very few of these anyway.

Replace PAPI_C with PAPI.  Most of the -dd messages in papi-c.c are
still PAPI.  There were only 3 uses of PAPI_C, all at init time.