Implement a sub-buildsystem for tests ("tests2")
The tests2 directory is Meson-based and can be configured with a previously
installed HPCToolkit. Includes a few simple tests, organized by component:
 - `tests2/hpcrun/`: Simple hpcrun test that operates on a test application with a single busy-loop
 - `tests2/hpcstruct/`: Simple hpcstruct test running on the simple single-loop binary
 - `tests2/hpcprof/`: Simple hpcprof test(s) that test processing works on small measurement inputs
 - `tests2/end2end/`: Simple end-to-end test running hpcrun/struct/prof on the simple binary

The hpcprof tests use measurements stored in the repository itself under `tests2/data/meas`.
An extra `make gen-testdata` target can be used to generate the data in this directory, for
ease of use (and proper data sanitization) a manual CI job is added at the end of the pipeline
to operate this target.

These tests are allowed to fail without stopping the CI pipeline. This allows us to still merge MRs and
continue rolling forward even when tests are failing. The MR widget and pipeline tab should make it clear
when there are failures that still need addressed.

A couple of vector-add GPU codes are included for demonstration purposes, they are not (yet) the subject
of any tests including HPCToolkit.

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/625