Fix for Issue 637.
If the instruction pointer in routine F in a leaf node maps to a
placeholder node P, rather than inserting a sibling node N representing
the first instruction F to record in the trace, simply record P in
the trace.

How this applies to issue 637:

When a sample is recorded in monitor_main, it is mapped to the placeholder
<program root>.  Instead of inserting a CCT node in tree for the first
instruction of `monitor_main` as a sibling of <program root> for recording
in the trace, simply insert <program root> in the trace.