Implement MetricScope::lex_aware
This new MetricScope is identical to MetricScope::function, except on loop Scopes
where it does not include child loops. This makes it the first (and only)
"lexically aware" MetricScope.

The meta.db lists `lex_aware` as a "custom" propagation scope, indicating that its
values are specially crafted and should not be tampered with. The metrics.yaml
now uses `lex_aware` for "exclusive:custom", and provides an "exclusive:standard"
fallback based on the (transitive) `function`.

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/631