hpcrun: eliminate cct modification race when pc sampling
long ago, we established the invariant that a GPU
monitoring thread could insert instruction PCs below a
GPU kernel placeholder. a requirement for this to be safe
was that no application thread could inspect or modify the
children of a GPU kernel placeholder. the function
gpu_cct_insert inserts a kernel_ip below a placeholder
once and returns no value.  when gpu_cct_insert finds that
a placeholder already has one or more children, it doesn't
try to insert a kernel_ip again or even look up the child
containing kernel_ip because it MUST already be there.
since a GPU monitoring thread may be inserting PCs
and rotating the splay tree of children, STAY CLEAR!

when adding support for level 0, this invariant was destroyed.
this commit separates the support for level 0 from gpu_cct_insert,
which preserves the "only modify children once" invariant into
a separate function gpu_cct_insert_always, which can return a
value. gpu_cct_insert, which doesn't inspect or modify a node's
children if one or more child is already present, now returns
no value, as it did previously.