Fix the handling for "uncalled" paths in GPU context reconstruction
There are cases where a whole function is never called because it was only used
by an entry point that was never launched. Unfortunately the previous code here
did not handle that case well, and it tripped the assertion.

This commit replaces it with very similar logic, but slightly better documented
and with some extra handling to ensure an uncalled Group is never added to the
working set.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/643