hpcrun: Align the stack for the child shim
Some architectures assume the stack pointer is always reasonably aligned,
but our tricks to clone() the Fnbounds server sometimes would cause the
stack pointer of the child shim (which shares the parent's stack) to
be misaligned.

So far this has only caused issues on ARM when running inside a container.
I'm not sure why it only happens when container tech is enabled, but it's
an innocuous change otherwise so might as well just do it.