Don't output an empty trace.db
When Run produces a measurements with tracing disabled, Prof2 would still
output a trace line for each application thread... even if it was empty.
This makes it very difficult to determine when traces contain any useful
information, the Viewer in this case would present a confusingly white box
for a trace view.

This commit implements a hotfix where, if the trace.db contains no timepoints
in the end, the entire trace.db is deleted before Prof completes. This is a
very coarse fix and only works in the non-MPI case.

This commit also adjusts a quirk of Prof2, where all "blank" samples at the
start of the trace would be elided. Now the start is treated like all other
"blank" samples and only the first is retained. This shouldn't make a
difference, but it better matches old-Prof and so may resolve some tricky
edge cases.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/624