Fix an internal edge case where Metrics get allocated overlapping ids
If a metric has no StatisticPartials (not even :Sum) the code would allocate overlapping ids to
the different Metrics. Also, the code only allocated enough space for the MetricScopes that
were used by the Metric, but always used 4 id-slots.

This patch rearranges the allocation of Metric ids to match the expectations.