metrics.yaml: Expose metric visibility
This commit adds a flag to metrics.yaml format to indicate when a metric should be hidden by default
and require the user to enable it in the metric table. This might change when we start working on the
hierarchical metric presentation, but for now a single boolean works.

Also, metrics marked as "invisible" were being written to the metrics.yaml when they shouldn't. This
commit properly implements that case as well.