Hpcrun4: Clean up properly when metric table parsing fails
If the metric table parsing fails for any reason, the entire Hpcrun4 aborts processing
as it should. Unfortunately what it did not do was freeze the created Metrics, leading
to an assertion error later on down the line.

Use RAII to ensure the successful Metrics are frozen no matter how metric parsing concludes.