SparseDB: Handle the no-profiles case
If there are no profiles (i.e. only the summary profile) due to parsing failures,
SparseDB would crash with an EOF error since the summary profile's header was not
yet written before attempting to read it in preparation for cct.db.

Avoid this by ignoring the unwritten segment and letting the resulting zeros propagate.