hpcprof: Accept metric formulas from PC sampling
The formulas used for derived metrics for PC sampling were improperly handled by my
quickly written infix expression parser. They contain whitespace and function calls
(for min and max). This patch adds the appropriate handling to the formula parser.

I'm *this* close to converting all the formulas to reverse-Polish notation, but
this approach was faster to implement.