hpcrun: support for restricting monitoring
Add an option to hpcrun so that an MPI rank is only
monitored if it is specified.

This commit is to support GAMESS, which oversubscribes GPUs. By
restricting monitoring to certain ranks, this commit enables hpcrun
to use PC sampling in the presence of GPU oversubscription by ignoring
GAMESS data server ranks.  We only want to monitor ranks that use GPUs,
not data-server ranks.

Details:
- add hpcrun options -olr --only-local-ranks
- add support for inspecting local rank information provided
  by SLURM, LSF, OpenMPI, MPICH