Unexpected test failure is an error
Tests that should pass (`should_fail: false`) should fail the pipeline when they unexpectedly fail. Similarly,
tests that are expected to fail (`should_fail: true`) should fail the pipeline when they pass.
This is the behavior of Meson's test harness and so matches the exit code of `make installcheck`.

Integration tests represent the set of features we want to support. Test failure indicates that there are bugs
preventing those features from working. Thus, `should_fail: true` tests are the "known bugs" in the current
revision, even when changing branches. This makes it possible to easily distinguish whether your latest change
broke something or whether it was broken before, significantly easing using tests in development.

Conversely, if a `should_fail: true` test passes, then a bug has been resolved and it should be marked as
`should_fail: false` once again. This makes it easy to maintain the list of known bugs as bugs are squashed
through normal development.

The last case is "flaky" tests, which fail unpredictably. These are at best an annoyance, a "flaky failure"
wastes valuble developer cycles in confirming the failure. A flaky test is not a good or useful test, period.
As such, these should be removed/commented out completely when identified, or at the very least configured to
always fail (and then marked as `should_fail: true`).