Expected test failure is a failure (in JUnit reports)
GitLab has a nifty MR widget that presents the results of the test suite based on the
JUnit result logs. The issue is that marking tests as `should_fail: true` makes them
passes in the JUnit log, which makes the widget lie about how many test *actually* failed.

Fix this situation by adjusting the JUnit logs on the way out. This makes the widget
present the actual number of failures and only reports tests as "fixed" when they
really do pass.

The widget has no influence on the mergability of the MR, so it fills this purpose nicely.