Fix the REALTIME zero samples bug for recent glibc.
Part 1 -- revert Johnmc's patch from Feb 2018 (e39c9f17ed3b).
Timerid is an opaque type and in recent glibc, zero is a legal value,
so can't set it to NULL.

This much fixes the zero samples bug, but should expose the "drain the
swamp" problem.  However, that problem is extremely difficult to
reproduce.