Fix timeout failures in the tests2/ CI
The Meson test harness by default runs using as many threads as there
are cores on the machine. This is fine for testing on development machines,
but in CI this differs drastically from the cgroups CPU-time limits.

Adjust ci.buildfe to limit the number of threads as exactly nproc_max(),
i.e. the number of CPU core equivalents available to the process.