Support Clang (10 and 14)
LLVM/Clang is a widely used compiler framework, being the base for many development tools
(clang-tidy, IWYU, etc.) in addition to actual compilers such as Cray. For full portability
to new and interesting environments, we should support compiling with Clang in addition to
the classic GCC, even if our dependencies don't.

This patch contains the minor tweaks needed to support Clang 10 and 14. Clang supports all
of C++17 since 7, but 10 is the oldest in the CI images. Likewise 14 is the newest.

Note that this adds ***many*** warnings to be investigated at a later date.