Update pre-commit hooks
Dodges a bug in the latest versions of one of the hooks, which only is popping
up now because the cache got reset from infrastructure updates.

To make this issue easier to detect in the future, disable the pre-commit cache
in CI and always pull the freshest from PyPI.

Thanks to the updated pylint and friends, most suppressed errors have gone away
or have better supression flags, so those are updated as well.