ci: Allow older CUDA drivers to run buildmany +cuda
If the runner is equipped with Nvidia hardware and nvidia-container-runtime, the runtime
will detect that CUDA is requested by the image and fail the container due to driver
version mismatch. It only worked up to now because not all runners are equipped with one
or both of those items.

Fix this by explicitly disabling hardware access, which disables the runtime check.
Since these jobs *only* build, they never needed access to the hardware anyway.