Report compiler warnings/errors as code quality results
With the addition of Clang to CI, we now have >800 warnings that are being reported
in every CI pipeline. But that value only appears deep in the CI job logs, where
(quite frankly) no one looks when reviewing an MR (unless the pipeline crashes).

This MR extends the rudimentary parsing of the compiler output to also generate a
"code quality" report for GitLab. This report gets presented as a widget on the MR,
listing the number of compiler errors ("critial") and warnings ("major") as well as
the summary line, compared to the target branch.

This should (in theory) allow us to keep an eye on the warnings as they evolve and
make it that much more obvious when an MR introduces new warning-ridden code.