[1/2] Pass shellcheck for most everything
A few of the scripts we install don't actually pass shellcheck, which is an issue
because the shellcheck linter is very "loose" (like the language itself), focusing
on specific counter-intuitive cases instead of understanding all of shell.
It has no false positives, only false negatives.

In other words, if shellcheck doesn't pass, you **definitely** do not have a good
shell script!

Autotools doesn't pass shellcheck. Which is a really bad sign, but proceeding anyway.