Do not output cct nodes with no metrics
We walked cct in node first and child next order to count number of nonzero cct nodes before,
now, since we are already checking if it has non-zero value or not, we just use this information
and also if it's been marked retained to decide if we want to output it in hpcrun file. Also, we
change the order to child first so that if child is being kept, then parent is automatically being
kept. Notice we don't change the actual writing order, this is just another walking before the
actual write. During the writing, if it's been marked display==true, then we output it.