Detect a logical exit based on any of the last 4 callers
Sometimes we will fail to match a logical region against a physical stack because
we've recorded the "wrong" exit frame to match on, because in Python we register
a callback instead of wrapping a function. So the stack frames look something like this:

    .-Outside Python caller
    |
    |         X-->X-->o  <--- We're here
    V        /
    o-->o-->o  <--- The frame want to record
             \
              o-->o  <--- Outside Python callee

The problem is we don't know how many X's there are, it changes between Python
builds/versions. We can't overestimate, or else we might "gobble" some caller
frames that are outside Python, and underestimating just doesn't work.

The solution in this commit: record multiple possibilities and pick the match
with fewest X's. So long as Python gives us our exit callback at the appropriate
times (it should), and there's no more than a small constant number of X's to skip
(in this commit 4) we won't gobble any of the caller frames by taking this approach.
This (should) ease the desync issues caused with other Python versions/builds.

There's still an issue on the beforeenter side. If Python -> C -> Python within
libpython.so (as happens during the init sequence), only the topmost logical region
will match, and it will match the entire call trace. That's not quite right, but
still figuring out how to handle that one.