Adjust warnings when using DWARF/ELF fallbacks
We want everyone to use Structfiles moving forward, they give better quality results
and we can make them faster than DWARF/ELF parsing will ever be. Encourage this by
emitting warnings when the DWARLF/ELF fallback is being used or when Structfiles are
simply not present.

As a side-effect, userdata for all objects is now lazily initialized. This reduces
the amount of parsing actually done (i.e. now only when needed), but also delays
it and may cause serialization that wasn't present before. Performance impact
in practice not yet investigated.

Structfiles are now considered perfect: if we have one, we trust that it is correct
even when it has a "gap." This is a little dangerous without some verification to
ensure that the Structfile has not gone stale.
See https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/651