Only ever borrow references to Python stack objects
Two of the Python functions (GetCode and GetBack) return "strong" references, which
increment the reference counter for the returned object. These counters are never
decremented which causes memory leaks of FrameObjects and local variables, etc.

Fix this by decrementing the count ourselves right after these calls. Since the Python
stack itself holds a "strong" reference to these objects, and because we get a
callback whenever the Python stack changes, we can just "borrow" that reference instead
of maintaining one a "strong" reference of our own. Probably.