hpcprof: no error when a callee of a callsite is not a function
output a warning when -vv is used, but don't consider this an error.
this case occurs when the callee is an internal function that
isn't considered as a first-class procedure. we observe this in
NVIDIA GPU kernels where there is a call to outlined code that
is actually surrounded by the calling procedure. in this case,
little information will be lost by just folding the cost into the
enclosing function since measurements are flat samples anyway.