hpcprof: Only trust/copy source files via absolute paths
If the File's path() was relative, the path will be resolved based on the
directory where hpcprof was run, NOT where the file was compiled. If a file
happens to exist at that relative path, it would then be copied (including
the relative path) into the database, which would often be wrong.

Fix this by assert()'ing that resolved paths are absolute, and refusing to
resolve paths that start out relative.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/681