Remove beforeenter_fixup, eagerly unwind out of libpython.so
Previously, beforeenter would be set to a caller of the callback function, to
be fully unwound lazily once a sample is taken. This of course only works if
the frame stored is "stable" and does not return before a sample or the paired
PyTrace_RETURN out of Python.

This commit removes the option for lazy unwinding, instead relying on eager
unwinding out of libpython.so. This is only needed once per call to Python
from C, so hopefully the increased overhead is not limiting.