Make the logical Python support opt-in
Currently there are version/ABI requirements that need to be met on
the user side, and it's unlikely those will be going away anytime soon.
To prevent surprises where an application crashes for strange reasons,
disable the Python support by default and only turn it on when requested.

This commit adds a new -a,--attribution super-option for settings that
alter how unwinds are done, passing `hpcrun -a python` enables the Python
support. For consistency, `-a flat` is an alternative name for -nu,--no-unwind.

This also removes the -a,--agent option used by LUSH. It's not documented
and IIRC LUSH is more-or-less unused, so I hope that's fine.