Enforce Python >= 3.10 in configure
3.10 is an overestimate. My comments indicate that the method should work for 3.8.
The original iteration was developed under 3.9. In theory it should work with any
version the `logical/python.c` file compiles under.

BUT the CI can only test >= 3.10, because I decided the testing infrastructure
should be written in Python 3.10, and I would rather not manage two different
Pythons at the same time. At least, not before we move to Meson and I can do that
dance without giving myself a headache.

So for now, Python >= 3.10 only.