Adjust hpcstruct encouragement warnings
There are cases where we expect to parse DWARF/ELF instead of using Structfiles, e.g. libhpcrun.so.
In these cases we don't want to issue a WARNING about parsing DWARF/ELF, to prevent confusing the
user by requesting them to repeat the operation they have already completed.

For full generality, this commit removes the WARNING when DWARF/ELF is being parsed. This removes
the confusion, since now the only WARNING we give to encourage hpcstruct use is when we detect that
an input measurements dir has not been "structified" yet, i.e. <measurements>/structs does not
exist yet. (Note that this logic is also slightly different, previously passing `-S` would suppress
this WARNING. This was changed to ensure "old style" users of Toolkit would still be notified that
a newer, more convenient alternative is available.)