feat: Support NDEBUG (replace assert() with abort())
`assert()` is a special macro in that it becomes a no-op whenever
`NDEBUG` is defined, i.e. in release builds. This makes it unsuitable
for error handling of any kind, it should only be used for internal
consistency and similar checks. Unfortunately, there are multiple cases
in the current codebase where `assert()` is used incorrectly.

To support `NDEBUG` in our codebase with minimal surprises, this commit
replaces most usage of `assert()` with conditional calls to `abort()`.
Specifically:

  - `assert(EXPR)` -> `if(!EXPR) abort()`,
  - `assert(0)` -> `abort()`, and
  - `assert(0 && "MSG")` -> `assert(false && "MSG"); abort()`.

This commit is primarily a mechanical change. No effort is made in this
commit to decide whether an assertion represents an internal consistency
check or not, all assertions are transformed regardless of their
context. Minimal changes are included as part of this commit to avoid
introducing new warnings, but that is all.

This commit does not touch the hpcprof or hpcserver code. The former
because it has followed this pattern for a long while, the latter
because hpcserver is and will remain unsupported.

The hpcrun code uses a slightly different transformation, a new function
`hpcrun_terminate()` is used in place of `abort()`. This function calls
`monitor_real_abort()` in most cases, however if `HPCRUN_ABORT_LIBC` is
set it calls `abort()` instead. This allows collecting callstacks while
under an MPI that intercepts `abort()` but does not collect coredumps.
For consistency, all other calls in hpcrun to `abort()` or
`monitor_real_abort()` have also been replaced by `hpcrun_terminate()`,
with the exception of the C++ and auditor shims.

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/718