build: Disable assertions in release builds
Many build systems, including Meson, consider a "release" build to be as
small and fast as possible. It is recommended to define `NDEBUG` in
these builds, which disables any and all `assert()` statements.

Enable `NDEBUG` in release builds. For Autotools-based builds, this
means with `--disable-debug`. For Meson-ic builds, this means with
`-Dbuildtype=release` following Meson's logic and the defined defaults.