hpcprof: Increase strictness of call graph handling
Previously, if a Structfile contained no <C> tags, it was assumed that the user
had run with --gpucfg yes and there happened to be no calls (i.e. a call graph
with no call edges). If the user ran with --gpucfg no, we would not emit an ERROR
about the missing CFG, and instead generate a very flat reconstructed Context tree.

With the addition of the <LM has-calls> attribute, we now only make the above assumption
if has-calls="1". If has-calls="0" (or not present), we instead consider a Structfile
with no <C> tags as indeed missing a call graph and report an ERROR.

This commit also re-enables the check for CFG validity. If the CFG contains an
invalid <C> tag, the CFG as a whole is considered erroneous and is wiped.