hpcprof: Lower Structfile-related ERRORs to WARNINGs
If a Structfile fails to parse, the (almost) worst that can happen is that we don't
get to classify some instructions with lexical Contexts (functions, loops, lines).
This does not imply that the resulting data will be wrong, just that the attribution
won't be as precise as we would otherwise like.
So failure here should be a WARNING, not an ERROR.

The other possible happening is that we end up with a missing CFG, which is reported
on the other end (Hpcrun4). This seems to cause metrics to attribute erroneously,
and we don't really want folks to believe our flat samples anyway ATM. So for now
this remains as an ERROR (although with a slightly better message now).

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/690