tests2: Add a correctness test for hpcstruct
By compiling a simple C program where we *know* what we should see in
the result, and matching for that pretty much exactly. The program is
designed so that the relavent structure cannot be optimized out by any
standards-compliant C compiler. Probably.

The pattern is coded in comment-pragmas in the source itself, in a cryptic
and undocumented pattern language. Something like XML mixed with UNIX globs.
Takes a PEG parser to understand. To be honest I'm surprised it works at all.

Very overcomplicated but best future is to merge with https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/694