ci: Redesign regen testdata jobs
Each `regen testdata` job regenerates a subset of the test data, so we
can't expect the test suite to pass until the final output is constructed.
On the flip side, we want to make sure that the test suite is not so
unstable that a single regeneration will cause it to fail.

This commit refactors the `regen testdata` job-sequence to only implement
regenerating testdata and not running any tests on it. This also moves the
`regen testdata` jobs into a child pipeline with a manual `trigger`,
ensuring that we only regenerate the test data when needed and don't slow
down pipelines because of it. Hopefully easier to use than MR labels.

The original functionality is still of interest to continually test the
test data for reproducability and stability, so this commit also adds a
new `check semifresh` job series (in the `check` family) to do just that.

Also included is some cleanup, moving some of the core CI pieces into ci.buildfe
and ci/scripts to share across the pipelines.

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/727