hpcrun: Properly handle signal safety for libunwind architectures (ARM)
`hpcrun_context_pc` is not async-signal-safe on ARM, but it needs to be called
to pass the PC to `hpcrun_safe_enter_async` (to check if the sample is in a trampoline).
This catch-22 causes synchronous unwinds (e.g. Python or GPU support) to fail for ARM.

This commit refactors `hpcrun_context_pc` into `hpcrun_context_pc_async`, which
is guarenteed to be async-signal-safe. The implementation is the same except for
the libunwind-based implementation, which always returns `NULL`.

Most prior usage of `hpcrun_context_pc` was for `hpcrun_safe_enter_async`, so the latter
is also refactored to take the context directly and internally call
`hpcrun_context_pc_async` (checking for `NULL` as appropriate). All other usage was
for log message, to reduce confusion these parts of the messages have been elided.

This is all correct because we can't use trampolines on ARM (AFAIK we would need custom
unwinder code), so if `hpcrun_context_pc_async` returns `NULL` we can't be a trampoline.

Fixes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/698.