Redesign how tests and test scripts are written
The old pattern (if there was one) made it difficult to distinguish between the test
script and the inputs, and had cryptic duplicate info in the test results line.

This commit rearranges the bits so that:
  - Every test script starts with a `tst-*` prefix, indicating that it is a test,
  - Every test script contains a (brief) Pythonic docstring documenting the test,
  - Every test is in suite(s) indicating what is being tested (i.e. tool names), and
  - Every test is named with an assertion describing what the test actually tests,
    that does NOT include the tool name.

The test scripts now use click to construct their basic CLIs, instead of the standard
but far less powerful argparse. In particular, click provides a natural syntax for
"sub-commands" which can be used later if some batch of tests share some bits in common
(but not enough to qualify for inclusion in the global hpctoolkit.test module).

There were also some tests that tested the inputs (executables). While this may
be needed during debugging, it doesn't test the software and thus isn't a "test."
Those tests have been removed for consistency.