hpcprof: Redesign Structfile-related warnings (again)
Since they were triggering a bit too often still. Let's try this again.

Hpcrun4 no longer emits a WARNING if CFG data is missing, now Structfile
and ProfArgs detect the issue earlier and report a clearer WARNINGs based
on the more specific information available there.

Specifically, if a measurements dir has no Structfiles, it reports:

    WARNING: Program structure data is missing from the measurements directory, performance attribution will be degraded
      Consider running hpcstruct first:
        $ hpcstruct path/to/measurements

If a Structfile doesn't have CFG data (i.e. hpcstruct was run with `--gpucfg no`),
and the profiles contain Contexts in need of reconstruction (i.e. PC sampling data),
it reports:

    WARNING: Unable to reconstruct calling contexts (or loops) within GPU kernels
      To enable this reconstruction, please run:
        $ hpcstruct --gpucfg yes path/to/measurements

Finally, if the Structfile was generated with CFG data (`--gpucfg yes`) but the CFG
or Structfile was corrupt, hpcprof reports one of the following much shorter:

    WARNING: Control flow data is corrupt, disabling affected GPU context reconstruction
    WARNING: Control flow data is missing or corrupted, disabling affected GPU context reconstruction

depending on whether the corruption was a bad CFG or bad XML, respectively.