tests2/data: Revamp test data regeneration and handling (take 3)
Previously there were two targets, gen-testdata and install-testdata. The former
would run all the custom_targets needed to generate a fresh batch of test data,
the latter would additionally copy the results back to the source directory.
This approach worked but has a couple of glaring issues moving forward:
 - Meson really doesn't like building "to" the source directory, so previously
   a custom `cp` command was used to "trick" Meson into doing what was wanted.
   This wasn't clean but it did work... at the cost of layers of complex Meson.
 - There was no way to ensure that *all* test data was regenerated, since there's
   only a single "gen-testdata" target it would operate under a "best-effort" principle
   and generate what it could with vauge system requirements.

To solve the latter issue, this commit splits the test data (re)generation path into "suites,"
each of which (will) have strict and well-defined system and software requirements to complete.
By requiring that each piece of test data be part of exactly one suite, we can ensure that all
test data is generated (exactly once) across multiple jobs. The dependency chains in each suite
make use of disabler() so that the final target will be disabled (and ninja will thus error)
if dependency issues arise at configure time.

To solve the former issue and decomplicate the Meson, the test data is now produced
*as a tarball* in the build directory (specifically a final custom_target). This removes
the complications from calling `cp`, and with the help of fs.is_samepath the path-equivalence
enforced by the previous approach is still assert()'d. The ci.buildfe now supports a series of
`-a fresh-testdata-{suite}` Actions that build the appropriate target and copy the output back
to the logs/ directory, which the CI then picks up and repackages for ease of use.