tests2/data: Unpack tarballs at configure time
Ninja doesn't actually support directories as build targets, but we were using them
to represent unpacked tarballs anyway. It's a lot easier to make well-behaved if we
instead do the extraction at configure time. If the extraction fails (i.e. new tarball)
we ignore the error, give a `warning()` and let the tests that need it later fail.

Also, remove the "pickling" step that was done to speed up hpcprof tests. It is
bound to the hpctoolkit.formats non-stable API, and it's easier to handle the error
if all that happens is tarball extraction. Parsing a database is already quite fast.