tests2: Fix CUDA dependency handling for Meson
The tests2 suite should be using the same CUDA that the Autogoo has been
passed, but previously it wasn't quite right due to a missing PATH adjustment.

Also remove the usage of the 'unstable-cuda' module. It is unstable and
can crash Meson on newer CUDAs, and doesn't actually make our work much,
if any, easier in the long run.